New Feature: support Android without GKI (kernel version > 4.18)
